requirements.txt
ccxt>=4.0.0
pyyaml>=6.0
aiofiles>=23.0.0
aiocsv>=1.2.0
aiohttp>=3.8.0
questionary>=2.0.0
rich>=13.0.0